requirements.txt
streamlit==1.29.0
langchain==0.1.0
langchain-openai==0.0.2
chromadb==0.4.22
pypdf==3.17.4
python-dotenv==1.0.0
openai==1.6.1
tiktoken==0.5.2